requirements.txt
pyautogui==0.9.54
pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.13
